Enable Segment Reload Status for REALTIME-Consuming tables
