+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/indexsegment/mutable/MutableSegmentImpl.java
@@ -21,7 +21,6 @@ package org.apache.pinot.segment.local.indexsegment.mutable;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Sets;
 import it.unimi.dsi.fastutil.ints.IntArrays;
-import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -53,7 +52,6 @@ import org.apache.pinot.segment.local.realtime.impl.json.MutableJsonIndex;
 import org.apache.pinot.segment.local.realtime.impl.nullvalue.MutableNullValueVector;
 import org.apache.pinot.segment.local.segment.creator.impl.V1Constants;
 import org.apache.pinot.segment.local.segment.index.datasource.ImmutableDataSource;
-import org.apache.pinot.segment.local.segment.index.datasource.MutableDataSource;
 import org.apache.pinot.segment.local.segment.index.metadata.SegmentMetadataImpl;
 import org.apache.pinot.segment.local.segment.index.readers.ValidDocIndexReaderImpl;
 import org.apache.pinot.segment.local.segment.virtualcolumn.VirtualColumnContext;
@@ -68,8 +66,6 @@ import org.apache.pinot.segment.spi.MutableSegment;
 import org.apache.pinot.segment.spi.SegmentMetadata;
 import org.apache.pinot.segment.spi.datasource.DataSource;
 import org.apache.pinot.segment.spi.index.creator.H3IndexConfig;
-import org.apache.pinot.segment.spi.index.reader.BloomFilterReader;
-import org.apache.pinot.segment.spi.index.reader.InvertedIndexReader;
 import org.apache.pinot.segment.spi.index.reader.MutableDictionary;
 import org.apache.pinot.segment.spi.index.reader.MutableForwardIndex;
 import org.apache.pinot.segment.spi.index.reader.ValidDocIndexReader;
@@ -740,8 +736,12 @@ public class MutableSegmentImpl implements MutableSegment {
       return new ImmutableDataSource(virtualColumnProvider.buildMetadata(virtualColumnContext),
           virtualColumnProvider.buildColumnIndexContainer(virtualColumnContext));
     } else {
-      return _indexContainerMap.get(column).toDataSource();
+      return _indexContainerMap.get(column).toDataSource(_numDocsIndexed);
     }
+  }
+
+  public Map<String, IndexContainer> getIndexContainerMap() {
+    return _indexContainerMap;
   }
 
   @Override
@@ -1050,7 +1050,7 @@ public class MutableSegmentImpl implements MutableSegment {
 
   // NOTE: Okay for single-writer
   @SuppressWarnings("NonAtomicOperationOnVolatileField")
-  private static class NumValuesInfo {
+  static class NumValuesInfo {
     volatile int _numValues = 0;
     volatile int _maxNumValuesPerMVEntry = -1;
 
@@ -1064,117 +1064,4 @@ public class MutableSegmentImpl implements MutableSegment {
     }
   }
 
-  private class IndexContainer implements Closeable {
-    final FieldSpec _fieldSpec;
-    final PartitionFunction _partitionFunction;
-    final Set<Integer> _partitions;
-    final NumValuesInfo _numValuesInfo;
-    final MutableForwardIndex _forwardIndex;
-    final MutableDictionary _dictionary;
-    final RealtimeInvertedIndexReader _invertedIndex;
-    final InvertedIndexReader _rangeIndex;
-    final MutableH3Index _h3Index;
-    final RealtimeLuceneTextIndexReader _textIndex;
-    final boolean _enableFST;
-    final MutableJsonIndex _jsonIndex;
-    final BloomFilterReader _bloomFilter;
-    final MutableNullValueVector _nullValueVector;
-
-    volatile Comparable _minValue;
-    volatile Comparable _maxValue;
-
-    // Hold the dictionary id for the latest record
-    int _dictId = Integer.MIN_VALUE;
-    int[] _dictIds;
-
-    IndexContainer(FieldSpec fieldSpec, @Nullable PartitionFunction partitionFunction,
-        @Nullable Set<Integer> partitions, NumValuesInfo numValuesInfo, MutableForwardIndex forwardIndex,
-        @Nullable MutableDictionary dictionary, @Nullable RealtimeInvertedIndexReader invertedIndex,
-        @Nullable InvertedIndexReader rangeIndex, @Nullable RealtimeLuceneTextIndexReader textIndex, boolean enableFST,
-        @Nullable MutableJsonIndex jsonIndex, @Nullable MutableH3Index h3Index, @Nullable BloomFilterReader bloomFilter,
-        @Nullable MutableNullValueVector nullValueVector) {
-      _fieldSpec = fieldSpec;
-      _partitionFunction = partitionFunction;
-      _partitions = partitions;
-      _numValuesInfo = numValuesInfo;
-      _forwardIndex = forwardIndex;
-      _dictionary = dictionary;
-      _invertedIndex = invertedIndex;
-      _rangeIndex = rangeIndex;
-      _h3Index = h3Index;
-
-      _textIndex = textIndex;
-      _enableFST = enableFST;
-      _jsonIndex = jsonIndex;
-      _bloomFilter = bloomFilter;
-      _nullValueVector = nullValueVector;
-    }
-
-    DataSource toDataSource() {
-      return new MutableDataSource(_fieldSpec, _numDocsIndexed, _numValuesInfo._numValues,
-          _numValuesInfo._maxNumValuesPerMVEntry, _partitionFunction, _partitions, _minValue, _maxValue, _forwardIndex,
-          _dictionary, _invertedIndex, _rangeIndex, _textIndex, _enableFST, _jsonIndex, _h3Index, _bloomFilter,
-          _nullValueVector);
-    }
-
-    @Override
-    public void close() {
-      String column = _fieldSpec.getName();
-      try {
-        _forwardIndex.close();
-      } catch (Exception e) {
-        _logger.error("Caught exception while closing forward index for column: {}, continuing with error", column, e);
-      }
-      if (_dictionary != null) {
-        try {
-          _dictionary.close();
-        } catch (Exception e) {
-          _logger.error("Caught exception while closing dictionary for column: {}, continuing with error", column, e);
-        }
-      }
-      if (_invertedIndex != null) {
-        try {
-          _invertedIndex.close();
-        } catch (Exception e) {
-          _logger
-              .error("Caught exception while closing inverted index for column: {}, continuing with error", column, e);
-        }
-      }
-      if (_rangeIndex != null) {
-        try {
-          _rangeIndex.close();
-        } catch (Exception e) {
-          _logger.error("Caught exception while closing range index for column: {}, continuing with error", column, e);
-        }
-      }
-      if (_textIndex != null) {
-        try {
-          _textIndex.close();
-        } catch (Exception e) {
-          _logger.error("Caught exception while closing text index for column: {}, continuing with error", column, e);
-        }
-      }
-      if (_jsonIndex != null) {
-        try {
-          _jsonIndex.close();
-        } catch (Exception e) {
-          _logger.error("Caught exception while closing json index for column: {}, continuing with error", column, e);
-        }
-      }
-      if (_h3Index != null) {
-        try {
-          _h3Index.close();
-        } catch (Exception e) {
-          _logger.error("Caught exception while closing H3 index for column: {}, continuing with error", column, e);
-        }
-      }
-      if (_bloomFilter != null) {
-        try {
-          _bloomFilter.close();
-        } catch (Exception e) {
-          _logger.error("Caught exception while closing bloom filter for column: {}, continuing with error", column, e);
-        }
-      }
-    }
-  }
 }
